No early BufferOverflowException checking
This would only work if the first message (serial 0) already arrived.
If that's not the case, we'd have for that message to arrive and throw
the exception at that time.

That would make the code more complex and the behavior
non-deterministic. Now, the exception simply happens if the message is
complete and the buffer *actually* overflows during the merge process.